requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pandas==2.2.0
requests==2.32.0
beautifulsoup4==4.12.0
selenium==4.25.0
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.0
